♻️ refactor(player): move clampVolume to AudioPlayer domain utils
clampVolume returns 1 (HTMLAudioElement default) when given NaN —
that fallback is volume-domain policy, not generic numeric clamping.
Both consumers (Context/reducer, Provider/AudioPlayerStateProvider)
already live inside the AudioPlayer domain. Move the helper and its
test from src/utils/ to components/AudioPlayer/utils/ so generic
src/utils/ stays purely for cross-domain helpers.
